Create apiUtils.ts
