+++ frontend/src/features/Favorite/Favorite.tsx
@@ -1,7 +1,8 @@
 import { useState, useEffect } from "react"
 import { Link } from "react-router-dom"
 import { Session } from "@supabase/supabase-js"
-import { Recipe, getUserFavoritesApi, deleteUserFavoritesApi } from "@/utils/recipes"
+import { Recipe } from "@/utils/recipes"
+import { getUserFavoritesApi, deleteUserFavoritesApi } from "@/utils/apiUtils"
 
 interface Props {
   session: Session | null
