+++ frontend/src/utils/recipes.ts
@@ -1,21 +1,3 @@
-const API_BASE_URL = import.meta.env.VITE_API_ENDPOINT
-
-export const postSearchRecipesApi = (): string => {
-  return `${API_BASE_URL}/api/searchRecipes`
-}
-
-export const getUserFavoritesApi = (): string => {
-  return `${API_BASE_URL}/api/users/favorites`
-}
-
-export const postUserFavoritesApi = (): string => {
-  return `${API_BASE_URL}/api/users/favorites`
-}
-
-export const deleteUserFavoritesApi = (id: number): string => {
-  return `${API_BASE_URL}/api/users/favorites/${id}`
-}
-
 export type Recipe = {
   id: number
   recipeTitle: string
